Add files via upload
Here we upload the codes to simulate an ecosystem describing interactions between Nematodes "smelling" and eating bacterial colonies that grow over time in a finite space. 
We have 2 versions of the code, one with a lot of parameters and one with less parameters and some changes to the systems behavior.